Simple HUD, Simple Frontend, Simple AI
Added a simple HUD element, ostensibly for aiming the tank cannon.
Added a front end menu that you can start the game from and go back to by pressing the "back" button on a gamepad.
Added an empty new AI controller and hooked that up as a default for tanks that are dropped into the world.
Added an empty player controller with some simple logging to show that it is hooked up to a tank_pawn.