Screen Space De-Projection For Aiming Via Reticle
Added override for Tick to TankPlayerController.
Added AimTowardCrosshair stub. Eventually going to move Tank_Pawn parts to aim at the HUD crosshair.
Added GetSightRayHitLocation which uses a screen space de-projection to do a line trace and see if it hits the world. If the line trace hits, return true and populate OutHitLocation with the location of the hit.